Serailize VM creation to prevent starting with wrong network.
On VM creation, we currently fire off a set of property adjustments
and then (if desired) autostart the VM. Unfortunately, this can result
in the VM starting before all properties are set-- resulting in a VM
that potentially conncets to the wrong network backend.

This commit moves the VM start code to execute after each of the
property adjustments has completed.